requirements.txt
Pillow==8.1.2
pyTelegramBotAPI==4.1.1
